Undependent lengths (#15)
* Update bignbr.c

Add support for using numbers of unequal length.
Also the number of digits doesn't need to include the sign & eon anymore. (The len argument of bignbr_init)

* Update bignbr_test.c

Adjust the length of some numbers to ether test the compatibility with unequal length or reduce the amount of memory needed.

* Update README.md

Adjust the length of the numbers in the example.